+++ src/main/java/com/pdv/lalapan/dto/VendaItemDTO.java
@@ -2,5 +2,5 @@ package com.pdv.lalapan.dto;
 
 import java.math.BigDecimal;
 
-public record VendaItemDTO(Long itemId, Long produtoId, String nomeProduto, BigDecimal precoUnitario, int quantidade) {
+public record VendaItemDTO(Long itemId, Long produtoId, String nomeProduto, BigDecimal precoUnitario, BigDecimal quantidade) {
 }
